Inittial version
Initial upload of the n-queen multithread and hadoop trials.
The idea is the break the brute force implementation (backtrack) in a
way it can be solved by multiple threads instead of a single thread.
The hadoop implementation is a test to generate results for very big
boards try to solve the problem faster (this solution need to be redone)